src/api/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
mypy==1.7.1
mypy-extensions==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
ruff==0.3.2
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.10.0
uritemplate==4.1.1
